Create README.md
This submission represents my skills test for the front-end position at Coalition Technologies. The objective of this assessment is to evaluate my capability to transform an Adobe Photoshop file into a static, single-page web application. The primary goal is to convert the provided PSD into a responsive single-page website. This entails ensuring that the carousel seamlessly adapts to various screen sizes, and on mobile devices, the tabs should gracefully transition into an accordion layout for improved usability.